capstone/requirements/requirements-gpu.txt
h5py==2.6.0
matplotlib==2.0.0
numpy==1.12.0
scipy==0.18.1
tqdm==4.11.2
keras==2.0.2
pillow==4.0.0
ipykernel==4.6.1
tensorflow-gpu==1.0.0